requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.0
pytz==2024.1
python-dateutil==2.9.0.post0
google-cloud-aiplatform==1.70.0
traceloop-sdk==0.47.5